rename binary to 'ocrd', merge run and run_server, update setup.py
